Extract viewable activity into its own file
